Move stivale in separate headers. Making them easier to use when including "limine" has a submodule. Also renamed stivale2_hdr_tag_framebuffer to stivale2_header_tag_framebuffer. This make it more consistant with other declaration in stivale2.
