+++ stage2/protos/stivale.c
@@ -16,38 +16,7 @@
 #include <drivers/pic.h>
 #include <fs/file.h>
 #include <mm/vmm64.h>
-
-struct stivale_header {
-    uint64_t stack;
-    uint16_t flags;
-    uint16_t framebuffer_width;
-    uint16_t framebuffer_height;
-    uint16_t framebuffer_bpp;
-    uint64_t entry_point;
-} __attribute__((packed));
-
-struct stivale_module {
-    uint64_t begin;
-    uint64_t end;
-    char     string[128];
-    uint64_t next;
-} __attribute__((packed));
-
-struct stivale_struct {
-    uint64_t cmdline;
-    uint64_t memory_map_addr;
-    uint64_t memory_map_entries;
-    uint64_t framebuffer_addr;
-    uint16_t framebuffer_pitch;
-    uint16_t framebuffer_width;
-    uint16_t framebuffer_height;
-    uint16_t framebuffer_bpp;
-    uint64_t rsdp;
-    uint64_t module_count;
-    uint64_t modules;
-    uint64_t epoch;
-    uint64_t flags;       // bit 0: 1 if booted with BIOS, 0 if booted with UEFI
-} __attribute__((packed));
+#include <stivale/stivale.h>
 
 #define KASLR_SLIDE_BITMASK 0x03FFFF000u
